test(copyright): split detector parser internals tests
